bundle - run from local relative paths
