+++ code/cell.compile.web/src/common/types.ts
@@ -1,1 +1,2 @@
 export * from '../types';
+export { INpmPackageJson } from '@platform/types';
