+++ code/cell.compile.web/src/node/index.ts
@@ -1,1 +1,1 @@
-export * from './bundler';
+export * from './bundle';
